tmb branch: batch_edit.sh: add an item to the edit all files plan. subgraph_sparsity.hpp: Change Pupose -> Notation. reverse_any.omh: white space edit. sparsity.hpp: remove unecessary type conversion.

+++ cppad/local/subgraph/sparsity.hpp
@@ -92,7 +92,7 @@ void subgraph_sparsity(
 {
 	// check dimension assumptions
 	CPPAD_ASSERT_UNKNOWN(
-		size_t( dep_taddr.size() ) == sub_info.n_dep()
+		dep_taddr.size() == sub_info.n_dep()
 	);
 	CPPAD_ASSERT_UNKNOWN(
 		size_t(select_domain.size()) == sub_info.n_ind()
